+++ front/src/components/Elements/Button/TopButton.tsx
@@ -1,12 +1,11 @@
+import { useSignedInUser } from '@/features/auth'
 import type { FC } from 'react'
 import { memo } from 'react'
 import { useNavigate } from 'react-router-dom'
 
-import { useAuth } from '@/features/auth'
-
 export const TopButton: FC = memo(() => {
   const navigate = useNavigate()
-  const { isSignedIn } = useAuth()
+  const { isSignedIn } = useSignedInUser()
   const onClickHome = () => {
     if (isSignedIn) {
       navigate('/users/home')
