+++ front/src/components/Pages/Home.tsx
@@ -5,11 +5,11 @@ import { Navigate, Link, useNavigate } from 'react-router-dom'
 import { PrimaryButton } from '@/components/Elements/Button'
 import { Head } from '@/components/Head'
 import { Header } from '@/components/Layout'
-import { useAuth } from '@/features/auth'
+import { useSignedInUser } from '@/features/auth'
 import { GuestSignInButton } from '@/features/auth/components/atoms/GuestSignInButton'
 
 export const Home: FC = memo(() => {
-  const { isSignedIn } = useAuth()
+  const { isSignedIn } = useSignedInUser()
   const navigate = useNavigate()
 
   // ルートパスアクセス時にログイン済みならリダイレクト
