+++ front/src/features/auth/index.ts
@@ -1,2 +1,3 @@
 export * from './types'
 export * from './hooks/useAuth'
+export * from './hooks/useSignedInUser'
