+++ front/src/features/users/components/pages/UserPassword.tsx
@@ -10,7 +10,7 @@ import { NotificationMessage } from '@/components/Elements/Notification'
 import { FormContainer, FormMain, FormTitle } from '@/components/Form'
 import { Head } from '@/components/Head'
 import type { PasswordResetParams } from '@/features/auth'
-import { useAuth } from '@/features/auth'
+import { useSignedInUser, useAuth } from '@/features/auth'
 import { updateUser } from '@/features/users/api/updateUser'
 import { PasswordConfirmationInput } from '@/features/users/components/molecules/PasswordConfirmationInput'
 import { PasswordInput } from '@/features/users/components/molecules/PasswordInput'
@@ -20,7 +20,8 @@ import { useNotification } from '@/hooks/useNotification'
 import type { SubmitHandler } from 'react-hook-form'
 
 export const UserPassword: FC = () => {
-  const { signedInUser, authHeaders } = useAuth()
+  const { authHeaders } = useAuth()
+  const { signedInUser } = useSignedInUser()
   const { showMessage } = useMessage()
   const { notifications } = useNotification()
   const navigate = useNavigate()
