+++ front/src/router/ProtectedRoute.tsx
@@ -1,7 +1,7 @@
 import type { ReactElement } from 'react'
 import { Navigate, Outlet } from 'react-router-dom'
 
-import { useAuth } from '@/features/auth'
+import { useSignedInUser } from '@/features/auth/hooks/useSignedInUser'
 
 type Props = {
   redirectPath?: string
@@ -10,7 +10,7 @@ type Props = {
 
 export const ProtectedRoute = (props: Props) => {
   const { redirectPath = '/auth/signin', children } = props
-  const { isSignedIn } = useAuth()
+  const { isSignedIn } = useSignedInUser()
   if (!isSignedIn) {
     return <Navigate to={redirectPath} replace />
   }
